OpenAPI Release: Increase resilience when provided with invalid data
If a manual release is created with a `target_commitish` of `master`, it
will compare `HEAD` with `master`, which contains no changes and never
trigger a release.

In addition, some OpenAPI docs contain version numbers that the semver
library does not parse. We now sanitise these inputs before version
comparison